requirements.txt
# ✅ COMPATIBLE WITH PYTHON 3.11-3.13
# ThaiDotRun Internship Project Requirements

# CORE DATA SCIENCE
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# VISUALIZATION
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# MACHINE LEARNING
scikit-learn==1.3.2
statsmodels==0.14.1

# JUPYTER & DEVELOPMENT
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1
ipywidgets==8.1.2

# DATA PROCESSING
openpyxl==3.1.2
pyarrow==14.0.1

# DASHBOARD
streamlit==1.29.0

# UTILITIES
python-dotenv==1.0.0
tqdm==4.66.1
pyyaml==6.0.1
pillow==10.1.0
joblib==1.3.2

# DATA PROFILING (REPLACEMENT for pandas-profiling)
ydata-profiling==4.6.1  # ✅ Works with Python 3.13

# MISSING DATA VISUALIZATION
missingno==0.5.2

# CODE QUALITY
black==23.11.0
flake8==6.1.0

# REMOVE THESE (incompatible):
# pandas-profiling==3.6.6  # ❌ DEPRECATED
# jupyter-contrib-nbextensions  # ❌ Causes issues
# xgboost  # ❌ Optional, install later if needed